requirements.txt
pandas
beautifulsoup4
matplotlib
seaborn
numpy
pyyaml
jinja2